Added charmverse space member property profile component
