+++ components/common/UserProfile/components/ProfileWidgets/ProfileWidgets.tsx
@@ -2,9 +2,10 @@ import { Grid } from '@mui/material';
 
 import { CollectionWidget } from './CollectionWidget';
 import { SocialWidget } from './SocialWidget';
+import { SpaceMemberPropertyWidget } from './SpaceMemberPropertyWidget';
 
 export function ProfileWidgets({ userId }: { userId: string }) {
-  const profileComponents = ['collection', 'ens', 'social', 'charmverse', 'lens'];
+  const profileComponents = ['collection', 'ens', 'social', 'charmverse', 'lens'] as const;
 
   return (
     <Grid container spacing={4}>
@@ -12,14 +13,21 @@ export function ProfileWidgets({ userId }: { userId: string }) {
         switch (profileComponent) {
           case 'collection':
             return (
-              <Grid item xs={12} md={6}>
+              <Grid item xs={12} md={6} alignItems='stretch'>
                 <CollectionWidget userId={userId} />
               </Grid>
             );
           case 'social':
             return (
-              <Grid item xs={12} md={6}>
+              <Grid item xs={12} md={6} alignItems='stretch'>
                 <SocialWidget userId={userId} />
+              </Grid>
+            );
+
+          case 'charmverse':
+            return (
+              <Grid item xs={12} md={6} alignItems='stretch'>
+                <SpaceMemberPropertyWidget userId={userId} />
               </Grid>
             );
           default:
